Added overrideNodeModulesPath config option
overrideNodeModulesPath allows you to change the nodeModulesPath that eslint will be loaded from. This is needed to workaround an issue using jest-runner-eslint within a yarn workspace. When using Yarn Workspaces the node modules are hoisted into the parent project's node_modules folder so eslint cannot be found in the "rootDir".